requirement.txt
streamlit
pandas
folium
streamlit-folium
geopandas
lightgbm
shapely
google-generativeai
ultralytics
opencv-python-headless
